STOCKS-115: Implement scaled unit management

+++ android-client/app/src/main/java/de/njsm/stocks/android/db/views/FoodItemView.java
@@ -25,6 +25,8 @@ import org.threeten.bp.Instant;
 
 import de.njsm.stocks.android.db.entities.VersionedData;
 
+import java.math.BigDecimal;
+
 public class FoodItemView extends VersionedData {
 
     private final String userName;
@@ -41,11 +43,11 @@ public class FoodItemView extends VersionedData {
 
     private final int scaledUnit;
 
-    private final int scale;
+    private final BigDecimal scale;
 
     private final String unitAbbreviation;
 
-    public FoodItemView(int id, @NonNull Instant validTimeStart, @NonNull Instant validTimeEnd, @NonNull Instant transactionTimeStart, @NonNull Instant transactionTimeEnd, int version, int initiates, String userName, String deviceName, Instant eatByDate, String location, int ofType, int storedIn, int scaledUnit, int scale, String unitAbbreviation) {
+    public FoodItemView(int id, @NonNull Instant validTimeStart, @NonNull Instant validTimeEnd, @NonNull Instant transactionTimeStart, @NonNull Instant transactionTimeEnd, int version, int initiates, String userName, String deviceName, Instant eatByDate, String location, int ofType, int storedIn, int scaledUnit, BigDecimal scale, String unitAbbreviation) {
         super(id, validTimeStart, validTimeEnd, transactionTimeStart, transactionTimeEnd, version, initiates);
         this.userName = userName;
         this.deviceName = deviceName;
@@ -86,7 +88,7 @@ public class FoodItemView extends VersionedData {
         return scaledUnit;
     }
 
-    public int getScale() {
+    public BigDecimal getScale() {
         return scale;
     }
